A few updates for a clean start
1. I don't think the changes in the cost context is relevant
2. We don't need a Ped in inner hash join. The new request (replicated outer
table) can be applied to outer joins as well. We can just add the new
optimization request to hash join.
3. In hash join, the optimization order is Right->Left. This isn't particularly
convenient for our new request, cause we need to be extra careful about
duplication hazard.
4. For now, don't worry about conditionally enforcing the replicated spec.
5. For now, disable commutativity transform.
6. We need to think about the usage of CEnfdDistribution::EdmSatisfy. For hash
join we request Satisfy for the first (inner) child and Exact for the second
(outer) child, with a few exceptions. We cannot change the Right->Left
optimization order. This means we will be request Satisfy for both children.
Need to think about what can go wrong.